Ändert die Addieren Funktion von int auf float

+++ Program.cs
@@ -12,11 +12,11 @@ namespace Taschenrechner_Iteration_1
       Console.Write("Bitte gebe den zweiten Summanden ein: ");
       string zweiterSummand = Console.ReadLine();
 
-      //  Wandelt Text in Ganzzahl um
-      int ersterSummandAlsZahl = Convert.ToInt32(ersterSummand);
-      int zweiterSummandAlsZahl = Convert.ToInt32(zweiterSummand);
+      //  Wandelt Text in Gleitkommazahlen um
+      float ersterSummandAlsZahl = Convert.ToSingle(ersterSummand);
+      float zweiterSummandAlsZahl = Convert.ToSingle(zweiterSummand);
       //  Berechnung ausführen
-      int summe = ersterSummandAlsZahl + zweiterSummandAlsZahl;
+      float summe = ersterSummandAlsZahl + zweiterSummandAlsZahl;
 
       //  Ausgabe
       Console.WriteLine("Die Summe ist: {0}", summe);
